ecommerce/requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
importlib-resources==6.1.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==23.2
pkgutil-resolve-name==1.3.10
pluggy==1.4.0
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.10.0
uritemplate==4.1.1
zipp==3.17.0
